CBD-2612: Remove all git submodule projects from couchbase-lite-* manifests
Also disable all couchbase-lite-* builds for cobalt and earlier releases.

We've updated the build-from-manifest script to handle git submodules without
needing to back-patch the input manifest, so now we want the input manifests
to list only the required projects and not all submodules. This will in some
cases show up as a "diff", which would trigger a new build, so I disabled
the older versions to prevent a flurry of unwanted builds.

Change-Id: Ie1f974571b1f3ff21115554c03112298db5ee2b6
Reviewed-on: http://review.couchbase.org/119641
Tested-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Blair Watt <blair.watt@couchbase.com>