Write a script to auto update all CBL manifests at once
This means that this particular commit, and future ones that touch a manifest for the first time, will be huge
because the order of the xml attributes are different and there is inconsistent spacing in the product-config.json
files.  However, further updates should be very small.  The main goal of this change is a version bump for all CBL
products to 2.6.4 for Cobalt and 2.7.1 for Mercury

Change-Id: I33040ed92274a52230986d1e8fa1f4ca017a4bc8
Reviewed-on: http://review.couchbase.org/121034
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Jim Borden <jim.borden@couchbase.com>