MB-40967: Bump HdrHistogram revision to fix int64->int32 conversion bug
hdr_iter takes the total_count of the histogram (h) at construction.
The count used in the histogram is a 64 bit count. When we have more
than std::numeric_limits<int32_t>::max() counts in the histogram the
total_count field of the hdr_iter will underflow when we create the
iterator and various >0 checks will check in and prevent iteration of
the histogram.

Change-Id: I222c697152c32ecd08945ce06698be2ed38eda2e
Reviewed-on: http://review.couchbase.org/c/manifest/+/134651
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>